Added gitignore, modified styles for submenu, written a new app.js file to bootstrap the app
App.js is now the main file for the application (I’ll try to put all
the new code inside it).
Modified the configuration file to be a simple Javascript one rather
than a json file; this was causing the app’s start up problem on the
device. Apparently, the json file didn’t exist locally on the device,
so the request was causing the problem.